Changes to Login and Gradle
User Registration now adds user object with login id to database.  Currently contains user's name and email address; this can be changed in the future.

Also made changes to Gradle to get it to compile; needed to add mulidex to it so that we didn't exceed the number of libraries we're allowed to call.