Fixed issue adding courses after registration/ booting app without logging in
Fixed an issue where the app would crash if the user tried to add a course after opening the app without logging in or if the user registered for the app but didn't login afterwards.  Also made my portion of the code in main just a tiny bit less horrible.